refactor: extract TinyModel to conftest, use valid_config_dict fixture
review-diff findings addressed:
- Extract duplicated TinyModel class from test_train_loop.py and
  test_degenerate_configs.py into conftest.py (shared test fixture)
- Replace 35-line inline config dict in test_falsify_01b with
  valid_config_dict fixture parameter (-61 lines)
