+++ tests/conftest.py
@@ -1,4 +1,6 @@
 import pytest
+import torch
+import torch.nn as nn
 
 # --- Minimum test count gate ---
 # Catches silent test collection failures (e.g., broken imports from
@@ -21,6 +23,31 @@ def pytest_collection_modifyitems(config, items):
                 f"expected at least {_MINIMUM_EXPECTED_TESTS}. "
                 f"Check for import errors in test files."
             )
+
+
+class TinyModel(nn.Module):
+    """
+    Minimal model matching HydraNet's interface contract:
+    forward(x, h) -> (reg_pred, cls_pred, h_next)
+
+    x: [B, C_in, H, W]
+    h: [B, hidden, H, W]
+    Returns: reg [B, n_reg, H, W], cls [B, n_cls, H, W], h [B, hidden, H, W]
+    """
+
+    def __init__(self, input_channels, n_reg, n_cls, hidden=4):
+        super().__init__()
+        self.base = hidden
+        self.reg_head = nn.Conv2d(input_channels + hidden, n_reg, 1)
+        self.cls_head = nn.Conv2d(input_channels + hidden, n_cls, 1)
+        self.h_update = nn.Conv2d(input_channels + hidden, hidden, 1)
+
+    def forward(self, x, h):
+        combined = torch.cat([x, h], dim=1)
+        return self.reg_head(combined), self.cls_head(combined), self.h_update(combined)
+
+    def init_hTtime(self, hidden_channels, H, W):
+        return torch.zeros(1, hidden_channels, H, W)
 
 
 @pytest.fixture
